Refactor gradle plugin classes to decouple access among class variables;
