+++ gradle_plugin/src/main/groovy/com/microsoft/hydralab/entity/AttachmentInfo.java
@@ -1,0 +1,9 @@
+package com.microsoft.hydralab.entity;
+
+public class AttachmentInfo {
+    public String fileName;
+    public String filePath;
+    public String fileType;
+    public String loadType;
+    public String loadDir;
+}
