+++ gradle_plugin/src/main/groovy/com/microsoft/hydralab/utils/HydraLabClientUtils.java
@@ -1,9 +1,11 @@
 package com.microsoft.hydralab.utils;
 
 import com.google.gson.*;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonWriter;
-import okhttp3.*;
+import com.microsoft.hydralab.entity.HydraLabAPIConfig;
+import com.microsoft.hydralab.entity.AttachmentInfo;
+import com.microsoft.hydralab.entity.BlobFileInfo;
+import com.microsoft.hydralab.entity.DeviceTestResult;
+import com.microsoft.hydralab.entity.TestTask;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
@@ -15,45 +17,15 @@ import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+
+import static com.microsoft.hydralab.utils.CommonUtils.*;
+import static com.microsoft.hydralab.utils.HydraLabAPIUtils.*;
 
 public class HydraLabClientUtils {
-    private static final OkHttpClient client = new OkHttpClient.Builder()
-            .readTimeout(300, TimeUnit.SECONDS)
-            .connectTimeout(300, TimeUnit.SECONDS)
-            .retryOnConnectionFailure(true)
-            .build();
 
     private static final int waitStartSec = 30;
     private static final int minWaitFinishSec = 15;
-    private static final int httpFailureRetryTimes = 10;
 
-    private static final Gson GSON = new GsonBuilder()
-            .registerTypeAdapter(Date.class, new TypeAdapter<Date>() {
-                @Override
-                public void write(JsonWriter out, Date value) throws IOException {
-                    if (value == null) {
-                        out.nullValue();
-                    } else {
-                        out.value(value.getTime());
-                    }
-                }
-
-                @Override
-                public Date read(JsonReader in) throws IOException {
-                    if (in != null) {
-                        try {
-                            return new Date(in.nextLong());
-                        } catch (IllegalStateException e) {
-                            in.nextNull();
-                            return null;
-                        }
-                    } else {
-                        return null;
-                    }
-                }
-            }).create();
     private static boolean isTestRunningFailed = false;
     private static boolean isTestResultFailed = false;
 
@@ -188,7 +160,7 @@ public class HydraLabClientUtils {
         }
 
         if (apiConfig == null) {
-            apiConfig = HydraLabAPIConfig.defaultAPI();
+            apiConfig = new HydraLabAPIConfig();
         }
         String testFileSetId = uploadApp(apiConfig, commitId, commitCount, commitMsg, app, testApp);
         printlnf("##[section]Uploaded test file set id: %s", testFileSetId);
@@ -440,41 +412,6 @@ public class HydraLabClientUtils {
         }
     }
 
-    private static void downloadToFile(String fileUrl, File file) {
-        Request req = new Request.Builder().get().url(fileUrl).build();
-        try (Response response = client.newCall(req).execute()) {
-            if (!response.isSuccessful()) {
-                return;
-            }
-            if (response.body() == null) {
-                return;
-            }
-            try (FileOutputStream fos = new FileOutputStream(file)) {
-                IOUtils.copy(response.body().byteStream(), fos);
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private static void assertNotNull(Object notnull, String argName) {
-        if (notnull == null) {
-            throw new IllegalArgumentException(argName + " is null");
-        }
-    }
-
-    private static void assertTrue(boolean beTrue, String msg, Object data) {
-        if (!beTrue) {
-            throw new IllegalStateException(msg + (data == null ? "" : ": " + data));
-        }
-    }
-
-    private static void printlnf(String format, Object... args) {
-        ZonedDateTime utc = ZonedDateTime.now(ZoneOffset.UTC);
-        System.out.print("[" + utc.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")) + "] ");
-        System.out.printf(format + "\n", args);
-    }
-
     private static void sleepIgnoreInterrupt(int second) {
         try {
             Thread.sleep(TimeUnit.SECONDS.toMillis(second));
@@ -483,359 +420,6 @@ public class HydraLabClientUtils {
         }
     }
 
-    private static void checkCenterAlive(HydraLabAPIConfig apiConfig) {
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(apiConfig.checkCenterAliveUrl())
-                .build();
-        OkHttpClient clientToUse = client;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Check center alive failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "check center alive", response);
-            printlnf("Center is alive, continue on requesting API...");
-        } catch (Exception e) {
-            throw new RuntimeException("check center alive fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static String uploadApp(HydraLabAPIConfig apiConfig, String commitId, String commitCount, String commitMsg, File app, File testApp) {
-        checkCenterAlive(apiConfig);
-
-        MediaType contentType = MediaType.get("application/vnd.android.package-archive");
-        MultipartBody.Builder multipartBodyBuilder = new MultipartBody.Builder()
-                .setType(MultipartBody.FORM)
-                .addFormDataPart("commitId", commitId)
-                .addFormDataPart("commitCount", commitCount)
-                .addFormDataPart("commitMessage", commitMsg)
-                .addFormDataPart("appFile", app.getName(), RequestBody.create(contentType, app));
-        if (!StringUtils.isEmpty(apiConfig.teamName)) {
-            multipartBodyBuilder.addFormDataPart("teamName", apiConfig.teamName);
-        }
-        if (testApp != null) {
-            multipartBodyBuilder.addFormDataPart("testAppFile", testApp.getName(), RequestBody.create(contentType, testApp));
-        }
-
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(apiConfig.getUploadUrl())
-                .post(multipartBodyBuilder.build())
-                .build();
-        OkHttpClient clientToUse = client;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Upload App failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "upload App", response);
-            ResponseBody body = response.body();
-
-            assertNotNull(body, response + ": upload App ResponseBody");
-            JsonObject jsonObject = GSON.fromJson(body.string(), JsonObject.class);
-
-            int resultCode = jsonObject.get("code").getAsInt();
-            assertTrue(resultCode == 200, "Server returned code: " + resultCode, jsonObject);
-
-            return jsonObject.getAsJsonObject("content").get("id").getAsString();
-        } catch (Exception e) {
-            throw new RuntimeException("upload App fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static JsonObject addAttachment(HydraLabAPIConfig apiConfig, String testFileSetId, AttachmentInfo attachmentConfig, File attachment) {
-        checkCenterAlive(apiConfig);
-
-        // default text file type: text/plain
-        // default binary file type: application/octet-stream
-        // todo: check if file is readable, set corresponding type
-        MediaType contentType = MediaType.get("application/octet-stream");
-        MultipartBody.Builder multipartBodyBuilder = new MultipartBody.Builder()
-                .setType(MultipartBody.FORM)
-                .addFormDataPart("fileSetId", testFileSetId)
-                .addFormDataPart("fileType", attachmentConfig.fileType)
-                .addFormDataPart("attachment", attachmentConfig.fileName, RequestBody.create(contentType, attachment));
-        if (StringUtils.isNotEmpty(attachmentConfig.loadType)) {
-            multipartBodyBuilder.addFormDataPart("loadType", attachmentConfig.loadType);
-        }
-        if (StringUtils.isNotEmpty(attachmentConfig.loadDir)) {
-            multipartBodyBuilder.addFormDataPart("loadDir", attachmentConfig.loadDir);
-        }
-
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(apiConfig.getAddAttachmentUrl())
-                .post(multipartBodyBuilder.build())
-                .build();
-        OkHttpClient clientToUse = client;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Add attachments failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "Add attachments", response);
-            ResponseBody body = response.body();
-            assertNotNull(body, response + ": Add attachments ResponseBody");
-
-            return GSON.fromJson(body.string(), JsonObject.class);
-        } catch (Exception e) {
-            throw new RuntimeException("Add attachments fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static String generateAccessKey(HydraLabAPIConfig apiConfig, String deviceIdentifier) {
-        checkCenterAlive(apiConfig);
-
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(String.format(apiConfig.getGenerateAccessKeyUrl(), deviceIdentifier))
-                .get()
-                .build();
-        OkHttpClient clientToUse = client;
-        JsonObject jsonObject = null;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Generate accessKey failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "generate accessKey", response);
-            ResponseBody body = response.body();
-
-            assertNotNull(body, response + ": generateAccessKey ResponseBody");
-            jsonObject = GSON.fromJson(body.string(), JsonObject.class);
-
-            int resultCode = jsonObject.get("code").getAsInt();
-            assertTrue(resultCode == 200, "Server returned code: " + resultCode, jsonObject);
-
-            return jsonObject.getAsJsonObject("content").get("key").getAsString();
-        } catch (Exception e) {
-            // TODO: no blocking for now, replace after enabling the access key usage
-            printlnf("##[warning]Request generateAccess failed: " + jsonObject.toString());
-            return "";
-//            throw new RuntimeException("generate accessKey fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static JsonObject triggerTestRun(String runningType, HydraLabAPIConfig apiConfig, String fileSetId, String testSuiteName,
-                                             String deviceIdentifier, @Nullable String accessKey, int runTimeoutSec, Map<String, String> instrumentationArgs, Map<String, String> extraArgs) {
-        checkCenterAlive(apiConfig);
-
-        JsonObject jsonElement = new JsonObject();
-        jsonElement.addProperty("runningType", runningType);
-        jsonElement.addProperty("deviceIdentifier", deviceIdentifier);
-        jsonElement.addProperty("fileSetId", fileSetId);
-        jsonElement.addProperty("testSuiteClass", testSuiteName);
-        jsonElement.addProperty("testTimeOutSec", runTimeoutSec);
-        jsonElement.addProperty("pkgName", apiConfig.pkgName);
-        jsonElement.addProperty("testPkgName", apiConfig.testPkgName);
-        jsonElement.addProperty("groupTestType", apiConfig.groupTestType);
-        jsonElement.addProperty("pipelineLink", apiConfig.pipelineLink);
-        jsonElement.addProperty("frameworkType", apiConfig.frameworkType);
-        jsonElement.addProperty("maxStepCount", apiConfig.maxStepCount);
-        jsonElement.addProperty("deviceTestCount", apiConfig.deviceTestCount);
-        jsonElement.addProperty("needUninstall", apiConfig.needUninstall);
-        jsonElement.addProperty("needClearData", apiConfig.needClearData);
-        jsonElement.addProperty("testRunnerName", apiConfig.testRunnerName);
-        jsonElement.addProperty("testScope", apiConfig.testScope);
-
-        try {
-            if (apiConfig.neededPermissions.size() > 0) {
-                jsonElement.add("neededPermissions", GSON.toJsonTree(apiConfig.neededPermissions));
-            }
-            if (StringUtils.isNotBlank(apiConfig.deviceActionsStr)) {
-                JsonParser parser = new JsonParser();
-                JsonObject jsonObject = parser.parse(apiConfig.deviceActionsStr).getAsJsonObject();
-                jsonElement.add("deviceActions", jsonObject);
-            }
-            if (instrumentationArgs != null) {
-                jsonElement.add("instrumentationArgs", GSON.toJsonTree(instrumentationArgs).getAsJsonObject());
-            }
-
-        } catch (JsonParseException e) {
-            throw new RuntimeException("trigger test running fail: " + e.getMessage(), e);
-        }
-
-        if (accessKey != null) {
-            jsonElement.addProperty("accessKey", accessKey);
-        }
-        if (extraArgs != null) {
-            extraArgs.forEach(jsonElement::addProperty);
-        }
-
-        String content = GSON.toJson(jsonElement);
-        printlnf("triggerTestRun api post body: %s", maskCred(content));
-        RequestBody jsonBody = RequestBody.create(MediaType.parse("application/json; charset=utf-8"), GSON.toJson(jsonElement));
-
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(apiConfig.getRunTestUrl())
-                .post(jsonBody).build();
-        OkHttpClient clientToUse = client;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Trigger test running failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "trigger test running", response);
-            ResponseBody body = response.body();
-            assertNotNull(body, response + ": triggerTestRun ResponseBody");
-            String string = body.string();
-            printlnf("RunningTestJson: %s", maskCred(string));
-            JsonObject jsonObject = GSON.fromJson(string, JsonObject.class);
-
-            return jsonObject;
-        } catch (Exception e) {
-            throw new RuntimeException("trigger test running fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static TestTask getTestStatus(HydraLabAPIConfig apiConfig, String testTaskId) {
-        checkCenterAlive(apiConfig);
-
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(apiConfig.getTestStatusUrl(testTaskId))
-                .build();
-        OkHttpClient clientToUse = client;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Get test status failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "get test status", response);
-            ResponseBody body = response.body();
-            assertNotNull(body, response + ": getTestStatus ResponseBody");
-            JsonObject jsonObject = GSON.fromJson(body.string(), JsonObject.class);
-
-            int resultCode = jsonObject.get("code").getAsInt();
-            assertTrue(resultCode == 200, "Server returned code: " + resultCode, jsonObject);
-
-            TestTask result = GSON.fromJson(jsonObject.getAsJsonObject("content"), TestTask.class);
-            if (result.id == null) {
-                result.id = testTaskId;
-            }
-            return result;
-        } catch (Exception e) {
-            throw new RuntimeException("get test status fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static String getBlobSAS(HydraLabAPIConfig apiConfig) {
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(apiConfig.getBlobSASUrl())
-                .build();
-        OkHttpClient clientToUse = client;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Get Blob SAS failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "Get Blob SAS", response);
-            ResponseBody body = response.body();
-
-            assertNotNull(body, response + ": Blob SAS");
-            JsonObject jsonObject = GSON.fromJson(body.string(), JsonObject.class);
-
-            int resultCode = jsonObject.get("code").getAsInt();
-            assertTrue(resultCode == 200, "Server returned code: " + resultCode, jsonObject);
-
-            return jsonObject.getAsJsonObject("content").get("signature").getAsString();
-        } catch (Exception e) {
-            throw new RuntimeException("Get Blob SAS fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static void cancelTestTask(HydraLabAPIConfig apiConfig, String testTaskId, String reason) {
-        checkCenterAlive(apiConfig);
-
-        Request req = new Request.Builder()
-                .addHeader("Authorization", "Bearer " + apiConfig.authToken)
-                .url(String.format(apiConfig.getCancelTestTaskUrl(), testTaskId, reason))
-                .build();
-        OkHttpClient clientToUse = client;
-        Response response = null;
-        try {
-            response = clientToUse.newCall(req).execute();
-            int waitingRetry = httpFailureRetryTimes;
-            while (!response.isSuccessful() && waitingRetry > 0) {
-                printlnf("##[warning]Cancel test task failed, remaining retry times: %d\nHttp code: %d\nHttp message: %s", waitingRetry, response.code(), response.message());
-                response = clientToUse.newCall(req).execute();
-                waitingRetry--;
-            }
-
-            assertTrue(response.isSuccessful(), "cancel test task", response);
-        } catch (Exception e) {
-            throw new RuntimeException("cancel test task fail: " + e.getMessage(), e);
-        } finally {
-            response.close();
-        }
-    }
-
-    private static String maskCred(String content) {
-        for (MaskSensitiveData sensitiveData : MaskSensitiveData.values()) {
-            Pattern PATTERNCARD = Pattern.compile(sensitiveData.getRegEx(), Pattern.CASE_INSENSITIVE);
-            Matcher matcher = PATTERNCARD.matcher(content);
-            if (matcher.find()) {
-                String maskedMessage = matcher.group(2);
-                if (maskedMessage.length() > 0) {
-                    content = content.replaceFirst(maskedMessage, "***");
-                }
-            }
-        }
-
-        return content;
-    }
 
     public static String getCommitCount(File commandDir, String startCommit) throws IOException {
         Process process = Runtime.getRuntime().exec(String.format("git rev-list --first-parent --right-only --count %s..HEAD", startCommit), null, commandDir.getAbsoluteFile());
@@ -861,210 +445,6 @@ public class HydraLabClientUtils {
             return IOUtils.toString(inputStream, StandardCharsets.UTF_8).trim();
         } finally {
             process.destroy();
-        }
-    }
-
-    // todo: split into APIConfig/deviceConfig/testConfig
-    public static class HydraLabAPIConfig {
-        public String schema = "https";
-        public String host = "";
-        public String contextPath = "";
-        public String authToken = "";
-        public boolean onlyAuthPost = true;
-        public String checkCenterVersionAPIPath = "/api/center/info";
-        public String checkCenterAliveAPIPath = "/api/center/isAlive";
-        public String getBlobSAS = "/api/package/getSAS";
-        public String uploadAPKAPIPath = "/api/package/add";
-        public String addAttachmentAPIPath = "/api/package/addAttachment";
-        public String generateAccessKeyAPIPath = "/api/deviceGroup/generate?deviceIdentifier=%s";
-        public String runTestAPIPath = "/api/test/task/run/";
-        public String testStatusAPIPath = "/api/test/task/";
-        public String cancelTestTaskAPIPath = "/api/test/task/cancel/%s?reason=%s";
-        public String testPortalTaskInfoPath = "/portal/index.html?redirectUrl=/info/task/";
-        public String testPortalTaskDeviceVideoPath = "/portal/index.html?redirectUrl=/info/videos/";
-        public String pkgName = "";
-        public String testPkgName = "";
-        public String groupTestType = "SINGLE";
-        public String pipelineLink = "";
-        public String frameworkType = "JUnit4";
-        public int maxStepCount = 100;
-        public int deviceTestCount = -1;
-        public boolean needUninstall = true;
-        public boolean needClearData = true;
-        public String teamName = "";
-        public String testRunnerName = "androidx.test.runner.AndroidJUnitRunner";
-        public String testScope = "";
-        public List<String> neededPermissions = new ArrayList<>();
-        public String deviceActionsStr = "";
-
-        public static HydraLabAPIConfig defaultAPI() {
-            return new HydraLabAPIConfig();
-        }
-
-        public String getBlobSASUrl() {
-            return String.format(Locale.US, "%s://%s%s%s", schema, host, contextPath, getBlobSAS);
-        }
-
-        public String checkCenterAliveUrl() {
-            return String.format(Locale.US, "%s://%s%s%s", schema, host, contextPath, checkCenterAliveAPIPath);
-        }
-
-        public String getUploadUrl() {
-            return String.format(Locale.US, "%s://%s%s%s", schema, host, contextPath, uploadAPKAPIPath);
-        }
-
-        public String getAddAttachmentUrl() {
-            return String.format(Locale.US, "%s://%s%s%s", schema, host, contextPath, addAttachmentAPIPath);
-        }
-
-        public String getGenerateAccessKeyUrl() {
-            return String.format(Locale.US, "%s://%s%s%s", schema, host, contextPath, generateAccessKeyAPIPath);
-        }
-
-        public String getRunTestUrl() {
-            return String.format(Locale.US, "%s://%s%s%s", schema, host, contextPath, runTestAPIPath);
-        }
-
-        public String getTestStatusUrl(String testTaskId) {
-            return String.format(Locale.US, "%s://%s%s%s%s", schema, host, contextPath, testStatusAPIPath, testTaskId);
-        }
-
-        public String getCancelTestTaskUrl() {
-            return String.format(Locale.US, "%s://%s%s%s", schema, host, contextPath, cancelTestTaskAPIPath);
-        }
-
-        public String getTestReportUrl(String testTaskId) {
-            return String.format(Locale.US, "%s://%s%s%s%s", schema, host, contextPath, testPortalTaskInfoPath, testTaskId);
-        }
-
-        public String getDeviceTestVideoUrl(String id) {
-            return String.format(Locale.US, "%s://%s%s%s%s", schema, host, contextPath, testPortalTaskDeviceVideoPath, id);
-        }
-
-        @Override
-        public String toString() {
-            return "HydraLabAPIConfig:\n" +
-                    "pkgName=" + pkgName + ",\n" +
-                    "testPkgName=" + testPkgName + ",\n" +
-                    "groupTestType=" + groupTestType + ",\n" +
-                    "pipelineLink=" + pipelineLink + ",\n" +
-                    "frameworkType=" + frameworkType + ",\n" +
-                    "maxStepCount=" + maxStepCount + ",\n" +
-                    "deviceTestCount=" + deviceTestCount + ",\n" +
-                    "needUninstall=" + needUninstall + ",\n" +
-                    "needClearData=" + needClearData + ",\n" +
-                    "teamName=" + teamName + ",\n" +
-                    "testRunnerName=" + testRunnerName + ",\n" +
-                    "testScope=" + testScope + ",\n" +
-                    "neededPermissions=" + (neededPermissions != null ? neededPermissions.toString() : "") + ",\n" +
-                    "deviceActionsStr=" + deviceActionsStr;
-        }
-    }
-
-    public static class AttachmentInfo {
-        String fileName;
-        String filePath;
-        String fileType;
-        String loadType;
-        String loadDir;
-    }
-
-    public static class TestTask {
-        public String id;
-        public List<DeviceTestResult> deviceTestResults;
-        public int testDevicesCount;
-        public Date startDate;
-        public Date endDate;
-        public int totalTestCount;
-        public int totalFailCount;
-        public String testSuite;
-        public String reportImagePath;
-        public String status;
-        public String testErrorMsg;
-        public String message;
-        public int retryTime;
-
-        @Override
-        public String toString() {
-            return "TestTask{" +
-                    "id='" + id + '\'' +
-                    ", testDevicesCount=" + testDevicesCount +
-                    ", startDate=" + startDate +
-                    ", totalTestCount=" + totalTestCount +
-                    ", status='" + status + '\'' +
-                    '}';
-        }
-
-        public interface TestStatus {
-            String RUNNING = "running";
-            String FINISHED = "finished";
-            String CANCELED = "canceled";
-            String EXCEPTION = "error";
-            String WAITING = "waiting";
-        }
-    }
-
-    public static class DeviceTestResult {
-        public String id;
-        public String deviceSerialNumber;
-        public String deviceName;
-        public String instrumentReportPath;
-        public String controlLogPath;
-        public String instrumentReportBlobUrl;
-        public String testXmlReportBlobUrl;
-        public String logcatBlobUrl;
-        public String testGifBlobUrl;
-
-        public List<BlobFileInfo> attachments;
-
-        public String crashStackId;
-        public String errorInProcess;
-
-        public String crashStack;
-
-        public int totalCount;
-        public int failCount;
-        public boolean success;
-        public long testStartTimeMillis;
-        public long testEndTimeMillis;
-
-        @Override
-        public String toString() {
-            return "{" +
-                    "SN='" + deviceSerialNumber + '\'' +
-                    ", totalCase:" + totalCount +
-                    ", failCase:" + failCount +
-                    ", success:" + success +
-                    '}';
-        }
-    }
-
-    public static class BlobFileInfo {
-        public String fileId;
-        public String fileType;
-        public String fileName;
-        public String blobUrl;
-        public String blobPath;
-        public long fileLen;
-        public String md5;
-        public String loadDir;
-        public String loadType;
-        public JsonObject fileParser;
-        public Date createTime;
-        public Date updateTime;
-
-
-        public interface fileType {
-            String WINDOWS_APP = "WINAPP";
-            String COMMOM_FILE = "COMMON";
-            String AGENT_PACKAGE = "PACKAGE";
-            String APP_FILE = "APP";
-            String TEST_APP_FILE = "TEST_APP";
-        }
-
-        public interface loadType {
-            String CPOY = "COPY";
-            String UNZIP = "UNZIP";
         }
     }
